requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
openai==1.3.3
